# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 知了 REST API
  version: 0.1.0
  description: |
    知了 MVP 的 HTTP 接口契约。房间内的麦位、排队、发言锁、公屏和在线状态
    使用 Socket.IO，详见 realtime-events.md。
servers:
  - url: /api/v1
    description: 与 Web 同源部署
...
paths:
...
  /auth/guest:
    post:
      tags: [Auth]
      operationId: createGuestSession
      summary: 创建或恢复游客会话
      description: |
        若请求已携带有效会话，则返回当前会话；否则创建本设备游客身份并写入
        HttpOnly Cookie。displayName 为空时由服务端生成随机昵称。
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 20
      responses:
        "200":
          description: 已有会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "201":
          description: 新建游客会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      summary: 获取当前会话
      responses:
        "200":
          description: 当前会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/zhihu/authorize:
    get:
      tags: [Auth]
      operationId: authorizeZhihu
      summary: 开始知乎 OAuth 授权
      description: 创建一次性 OAuth 尝试 Cookie 后跳转到知乎授权页。
      security: []
      parameters:
        - name: returnTo
          in: query
          required: false
          description: OAuth 成功后返回的站内路径，不接受外部 URL。
          schema:
            type: string
            maxLength: 500
            default: /
      responses:
        "302":
          description: 跳转知乎授权页
          headers:
            Location:
              schema:
                type: string
                format: uri
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamFailure"
  /auth/zhihu/callback:
    get:
      tags: [Auth]
      operationId: handleZhihuCallback
      summary: 知乎 OAuth 回调
      description: 服务端换取 Token，将游客会话升级为知乎授权会话后跳转回 Web。
      security: []
      parameters:
        - name: authorization_code
          in: query
          required: true
          description: 知乎 OAuth 回调返回的授权码。
          schema:
            type: string
        - name: state
          in: query
          required: false
          description: 若知乎授权服务回传该字段则必须校验。
          schema:
            type: string
      responses:
        "302":
          description: 授权完成后跳转站内 returnTo
          headers:
            Location:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamFailure"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: zhiliao_session
...
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 缺少或失效会话
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UpstreamFailure:
      description: 外部服务不可用
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    IdentityType:
      type: string
      enum: [guest, zhihu]
    UserLevel:
      type: integer
      enum: [1, 2, 3, 4, 5, 6]
    UserLevelTitle:
      type: string
      enum: [蛰伏, 破土, 蜕壳, 振翅, 鸣夏, 知秋]
    PublicUser:
      type: object
      additionalProperties: false
      required: [userId, identityType, displayName, avatarUrl, level, levelTitle]
      properties:
        userId:
          type: string
        identityType:
          $ref: "#/components/schemas/IdentityType"
        displayName:
          type: string
        avatarUrl:
          type: [string, "null"]
          format: uri
        level:
          $ref: "#/components/schemas/UserLevel"
        levelTitle:
          $ref: "#/components/schemas/UserLevelTitle"
    UserAccount:
      type: object
      additionalProperties: false
      required: [coinBalance, experience, level, levelTitle, nextLevelExperience]
      properties:
        coinBalance:
          type: integer
          minimum: 0
        experience:
          type: integer
          minimum: 0
        level:
          $ref: "#/components/schemas/UserLevel"
        levelTitle:
          $ref: "#/components/schemas/UserLevelTitle"
        nextLevelExperience:
          type: [integer, "null"]
          minimum: 0
    SessionResponse:
      type: object
      additionalProperties: false
      required: [user, account, permissions]
      properties:
        user:
          $ref: "#/components/schemas/PublicUser"
        account:
          $ref: "#/components/schemas/UserAccount"
        permissions:
          type: object
          additionalProperties: false
          required: [canCreateRoom, canRequestSeat, canSpeak]
          properties:
            canCreateRoom:
              type: boolean
            canRequestSeat:
              type: boolean
            canSpeak:
              type: boolean
...
    ErrorDetail:
      type: object
      additionalProperties: true
      required: [code, message, details, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        requestId:
          type: string
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
security:
  - cookieAuth: []